// src/game/projectileManifest.generated.ts
// AUTO-GENERATED by scripts/build-sheets.mjs — do not edit by hand.
export type ProjectileType = "DeathAoE" | "FireAoE" | "IceAoE" | "SwordAoE" | "Arrow" | "FireArrow" | "ArcSpell" | "DeathSpell" | "FireSpell" | "IceSpell";
export interface ProjectileMeta { category: string; frames: number[]; start: number; w: number; h: number; }

export const PROJECTILE_MANIFEST: Record<ProjectileType, ProjectileMeta> = {
  "DeathAoE": {
    "category": "AoE",
    "frames": [
      1,
      3,
      5,
      7,
      9,
      11,
      13,
      15,
      17,
      19,
      21,
      23,
      25,
      27,
      29
    ],
    "start": 1,
    "w": 512,
    "h": 512
  },
  "FireAoE": {
    "category": "AoE",
    "frames": [
      1,
      3,
      5,
      7,
      9,
      11,
      13,
      15,
      17,
      19,
      21,
      23,
      25,
      27,
      29
    ],
    "start": 4,
    "w": 512,
    "h": 512
  },
  "IceAoE": {
    "category": "AoE",
    "frames": [
      1,
      3,
      5,
      7,
      9,
      11,
      13,
      15,
      17,
      19,
      21,
      23,
      25,
      27,
      29
    ],
    "start": 6,
    "w": 512,
    "h": 512
  },
  "SwordAoE": {
    "category": "AoE",
    "frames": [
      1,
      3,
      5,
      7,
      9,
      11,
      13,
      15,
      17,
      19,
      21,
      23,
      25,
      27,
      29
    ],
    "start": 2,
    "w": 256,
    "h": 256
  },
  "Arrow": {
    "category": "Arrows",
    "frames": [
      1,
      3,
      5,
      7,
      9,
      11,
      13,
      15,
      17,
      19,
      21,
      23,
      25,
      27,
      29
    ],
    "start": 5,
    "w": 512,
    "h": 26
  },
  "FireArrow": {
    "category": "Arrows",
    "frames": [
      1,
      3,
      5,
      7,
      9,
      11,
      13,
      15,
      17,
      19,
      21,
      23,
      25,
      27,
      29
    ],
    "start": 2,
    "w": 512,
    "h": 26
  },
  "ArcSpell": {
    "category": "Spells",
    "frames": [
      1,
      3,
      5,
      7,
      9,
      11,
      13,
      15,
      17,
      19,
      21,
      23,
      25,
      27,
      29
    ],
    "start": 2,
    "w": 256,
    "h": 256
  },
  "DeathSpell": {
    "category": "Spells",
    "frames": [
      1,
      3,
      5,
      7,
      9,
      11,
      13,
      15,
      17,
      19,
      21,
      23,
      25,
      27,
      29
    ],
    "start": 1,
    "w": 256,
    "h": 64
  },
  "FireSpell": {
    "category": "Spells",
    "frames": [
      1,
      3,
      5,
      7,
      9,
      11,
      13,
      15,
      17,
      19,
      21,
      23,
      25,
      27,
      29
    ],
    "start": 2,
    "w": 256,
    "h": 64
  },
  "IceSpell": {
    "category": "Spells",
    "frames": [
      1,
      3,
      5,
      7,
      9,
      11,
      13,
      15,
      17,
      19,
      21,
      23,
      25,
      27,
      29
    ],
    "start": 1,
    "w": 256,
    "h": 64
  }
};